+++ src/Resources/contao/config/config.php
@@ -142,6 +142,7 @@ $GLOBALS['FE_MOD']['fernschachverwaltung'] = array
 	'fernschachverwaltung_titelnormen'          => 'Schachbulle\ContaoFernschachBundle\Modules\TitelNormen',
 	'fernschachverwaltung_titelnormen_liste'    => 'Schachbulle\ContaoFernschachBundle\Modules\TitelNormenLast',
 	'fernschachverwaltung_kontoauszug'          => 'Schachbulle\ContaoFernschachBundle\Modules\Kontoauszug',
+	'fernschachverwaltung_beitritt'             => 'Schachbulle\ContaoFernschachBundle\Modules\Beitrittserklaerung',
 );
 
 /*